Moved memstorage from main in repositories
