+++ cmd/server/main.go
@@ -4,9 +4,9 @@ import (
 	"fmt"
 	"net/http"
 	"strconv"
-	"sync"
 
 	"github.com/itaraxa/effectivepancake/internal/models"
+	"github.com/itaraxa/effectivepancake/internal/repositories/memstorage"
 )
 
 // Интерфейс для описания взаимодействия с хранилищем метрик
@@ -17,46 +17,10 @@ type Storager interface {
 	GetCounter(metric_name string) ([]int64, error)
 }
 
-// Структура для хранения метрик в памяти
-type MemStorage struct {
-	Gauge   map[string]float64
-	Counter map[string][]int64
-	mu      sync.Mutex
-}
-
-func (m *MemStorage) UpdateGauge(metric_name string, value float64) error {
-	return nil
-}
-
-func (m *MemStorage) AddCounter(metric_name string, value int64) error {
-	return nil
-}
-
-func (m *MemStorage) GetGauge(metric_name string) (float64, error) {
-	return 0.0, nil
-}
-
-func (m *MemStorage) GetCounter(metric_name string) ([]int64, error) {
-	return []int64{}, nil
-}
-
-func (m *MemStorage) String() string {
-	s := ">>>MemStorage<<<\n\r"
-	s += "Gauge:\n\r"
-	for metric, value := range m.Gauge {
-		s += fmt.Sprintf("  %s: %f\n\r", metric, value)
-	}
-	s += "Counter:\n\r"
-	for metric, values := range m.Counter {
-		s += fmt.Sprintf("  %s: %v\n\r", metric, values)
-	}
-	return s
-}
-
-func UpdateMemStorageHandler(ms *MemStorage) http.HandlerFunc {
+func UpdateMemStorageHandler(ms *memstorage.MemStorage) http.HandlerFunc {
 	return func(w http.ResponseWriter, req *http.Request) {
-		ms.mu.Lock()
-		defer ms.mu.Unlock()
+		// ms.mu.Lock()
+		// defer ms.mu.Unlock()
 
 		// Проверки запроса
 		if req.Method != http.MethodPost {
@@ -103,7 +67,7 @@ func UpdateMemStorageHandler(ms *MemStorage) http.HandlerFunc {
 }
 
 func main() {
-	ms := &MemStorage{
+	ms := &memstorage.MemStorage{
 		Gauge:   make(map[string]float64),
 		Counter: make(map[string][]int64),
 	}
